LUTs now range [0.0, 1.0]

+++ firmware/inc/lookup_tables.h
@@ -3,12 +3,10 @@
 #ifndef LOOKUP_TABLES_H_INCLUDED
 #define LOOKUP_TABLES_H_INCLUDED
 
-#include <stdint.h>
-
 // the attack curve lookup table
-extern const uint32_t ADSR_ATTACK_TABLE[1024];
+extern const float ADSR_ATTACK_TABLE[1024];
 
 // the decay curve lookup table, used for decay and release curves
-extern const uint32_t ADSR_DECAY_TABLE[1024];
+extern const float ADSR_DECAY_TABLE[1024];
 
 #endif
